Move FromConfigAdapter to config package
