+++ .eslintignore
@@ -6,4 +6,5 @@
 **/umd/**
 # TODO: decide whether or not to bring these up to eslint standards
 packages/jbrowse-web/scripts/**
-config/
+/config/
+/packages/jbrowse-web/config/
