+++ packages/config/src/ConfigurationEditorDrawerWidget/components/ColorEditor.js
