+++ packages/config/src/ConfigurationEditorDrawerWidget/components/ConfigurationEditor.js
