+++ packages/config/src/ConfigurationEditorDrawerWidget/components/ConfigurationEditor.test.js
