+++ packages/config/src/ConfigurationEditorDrawerWidget/components/JsonEditor.js
