+++ packages/config/src/ConfigurationEditorDrawerWidget/components/SlotEditor.js
