+++ packages/config/src/ConfigurationEditorDrawerWidget/components/TypeSelector.js
