+++ packages/config/src/ConfigurationEditorDrawerWidget/index.js
