+++ packages/config/src/ConfigurationEditorDrawerWidget/model.js
