+++ packages/config/src/FromConfigAdapter/FromConfigAdapter.ts
