+++ packages/config/src/FromConfigAdapter/index.js
@@ -1,0 +1,14 @@
+import { ConfigurationSchema } from '@gmod/jbrowse-core/configuration'
+
+export { default as AdapterClass } from './FromConfigAdapter'
+
+export const configSchema = ConfigurationSchema(
+  'FromConfigAdapter',
+  {
+    features: {
+      type: 'frozen',
+      defaultValue: [],
+    },
+  },
+  { explicitlyTyped: true },
+)
