+++ packages/config/src/FromConfigAdapter/test_data/volvox.2bit
