+++ packages/jbrowse-web/package.json
@@ -18,7 +18,7 @@
     "@gmod/jbrowse-core": "^2.0.0",
     "@gmod/jbrowse-plugin-alignments": "^2.0.0",
     "@gmod/jbrowse-plugin-bed": "^2.0.0",
-    "@gmod/jbrowse-plugin-config-editing": "^2.0.0",
+    "@gmod/jbrowse-plugin-config": "^2.0.0",
     "@gmod/jbrowse-plugin-data-management": "^2.0.0",
     "@gmod/jbrowse-plugin-jbrowse1": "^2.0.0",
     "@gmod/jbrowse-plugin-linear-genome-view": "^2.0.0",
