+++ packages/jbrowse-web/src/corePlugins.js
@@ -1,6 +1,6 @@
 import JBrowse1 from '@gmod/jbrowse-plugin-jbrowse1'
 import DataManagement from '@gmod/jbrowse-plugin-data-management'
-import ConfigEditing from '@gmod/jbrowse-plugin-config-editing'
+import Config from '@gmod/jbrowse-plugin-config'
 import TrackHubRegistry from '@gmod/jbrowse-plugin-trackhub-registry'
 import LinearGenomeView from '@gmod/jbrowse-plugin-linear-genome-view'
 import Alignments from '@gmod/jbrowse-plugin-alignments'
@@ -11,7 +11,6 @@ import Wiggle from '@gmod/jbrowse-plugin-wiggle'
 import MainMenuBarPlugin from './plugins/MainMenuBar'
 
 // adapters
-import FromConfigAdapterPlugin from './plugins/FromConfigAdapter'
 
 // tracks
 import FilteringTrackPlugin from './plugins/FilteringTrack'
@@ -24,12 +23,11 @@ export default [
   Alignments,
   Sequence,
   DataManagement,
-  ConfigEditing,
+  Config,
   MainMenuBarPlugin,
   Bed,
   LinearGenomeView,
   SvgFeaturePlugin,
-  FromConfigAdapterPlugin,
   FilteringTrackPlugin,
   LollipopRendererPlugin,
   JBrowse1,
