--- packages/jbrowse-web/src/plugins/FromConfigAdapter/index.js
@@ -1,29 +1,0 @@
-import Plugin from '@gmod/jbrowse-core/Plugin'
-import AdapterType from '@gmod/jbrowse-core/pluggableElementTypes/AdapterType'
-import { ConfigurationSchema } from '@gmod/jbrowse-core/configuration'
-
-import AdapterClass from './FromConfigAdapter'
-
-const configSchema = ConfigurationSchema(
-  'FromConfigAdapter',
-  {
-    features: {
-      type: 'frozen',
-      defaultValue: [],
-    },
-  },
-  { explicitlyTyped: true },
-)
-
-export default class FromConfigAdapterPlugin extends Plugin {
-  install(pluginManager) {
-    pluginManager.addAdapterType(
-      () =>
-        new AdapterType({
-          name: 'FromConfigAdapter',
-          configSchema,
-          AdapterClass,
-        }),
-    )
-  }
-}
