+++ packages/protein-widget/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@gmod/gff": "^1.1.2",
     "@gmod/jbrowse-core": "^2.0.0",
+    "@gmod/jbrowse-plugin-config": "^2.0.0",
     "@gmod/jbrowse-plugin-linear-genome-view": "^2.0.0",
     "@gmod/jbrowse-plugin-protein": "^2.0.0",
     "mobx": "^5.9.0",
