+++ packages/protein-widget/src/Viewer.js
@@ -13,14 +13,14 @@ import { ConfigurationSchema } from '@gmod/jbrowse-core/configuration'
 import RpcManager from '@gmod/jbrowse-core/rpc/RpcManager'
 
 import LinearGenomeView from '@gmod/jbrowse-plugin-linear-genome-view'
+import Config from '@gmod/jbrowse-plugin-config'
 import Protein from '@gmod/jbrowse-plugin-protein'
-import FromConfigAdapterPlugin from '../../jbrowse-web/src/plugins/FromConfigAdapter'
 import LollipopRendererPlugin from '../../jbrowse-web/src/plugins/LollipopRenderer'
 import SvgFeatureRendererPlugin from '../../jbrowse-web/src/plugins/SvgFeatureRenderer'
 import FilteringTrackPlugin from '../../jbrowse-web/src/plugins/FilteringTrack'
 
 const plugins = [
-  FromConfigAdapterPlugin,
+  Config,
   LinearGenomeView,
   Protein,
   LollipopRendererPlugin,
